requirements.txt
alembic==1.13.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx2txt==0.8
Flask==3.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
future==0.18.3
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.4
mime==0.1.0
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.1
